+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### v1.0.2 `2017-04-04`
 - Fix usage of single persona alias. Fixes [#1](https://github.com/raphaelstolt/git-user-bend/issues/1).
+- Additional guard against dual alias and aliases argument usage on `use` command.
 
 #### v1.0.1 `2017-04-03`
 - Some minor bug fixes.